+++ Hearthstone Deck Tracker/Windows/MainWindow.Export.cs
@@ -1,15 +1,12 @@
 ﻿#region
 
 using System;
-using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Forms;
 using System.Windows.Media.Imaging;
-using HearthMirror;
-using Hearthstone_Deck_Tracker.Exporting;
 using Hearthstone_Deck_Tracker.Hearthstone;
 using Hearthstone_Deck_Tracker.Utility.Extensions;
 using Hearthstone_Deck_Tracker.Utility.Logging;
@@ -24,76 +21,10 @@ namespace Hearthstone_Deck_Tracker.Windows
 {
 	public partial class MainWindow
 	{
-		internal async void ExportDeck(Deck deck)
+		public void ShowExportFlyout(Deck deck)
 		{
-			if(deck == null)
-				return;
-			if(Config.Instance.ShowExportingDialog)
-			{
-				var message = $"1) Create a new (or open an existing) {deck.Class} deck.\n\n2) Leave the deck creation screen open.\n\n3) Click 'Export' and do not move your mouse or type until done.";
-				var result = await this.ShowMessageAsync("Export " + deck.Name + " to Hearthstone", message, AffirmativeAndNegative, new MessageDialogs.Settings { AffirmativeButtonText = "Export" });
-				if(result == MessageDialogResult.Negative)
-					return;
-			}
-			HearthMirror.Objects.Deck openDeck;
-			var settings = new MessageDialogs.Settings() {AffirmativeButtonText = "Continue", NegativeButtonText = "Cancel"};
-			while((openDeck = Reflection.GetEditedDeck()) == null)
-			{
-				var result = await this.ShowMessageAsync("No open deck found", "Please open a deck for editing in Hearthstone before continuing.", AffirmativeAndNegative, settings);
-				if(result == MessageDialogResult.Negative)
-					return;
-			}
-			var missingCards = ExportingHelper.GetMissingCards(deck).ToList();
-			deck.MissingCards = missingCards;
-			if(missingCards.Count > 0)
-			{
-				var result = await this.ShowMissingCardsMessage(deck, true);
-				if(result == MessageDialogResult.Negative)
-					return;
-			}
-			string selectedClass;
-			while((selectedClass = Database.GetCardFromId(openDeck.Hero).PlayerClass) != deck.Class)
-			{
-				var result = await this.ShowMessageAsync("Incorrect class", $"Open deck is a {selectedClass} deck, but we are trying to import a {deck.Class} deck. Please create a deck with the correct class before continuing.", AffirmativeAndNegative, settings);
-				if(result == MessageDialogResult.Negative)
-					return;
-				openDeck = Reflection.GetEditedDeck();
-			}
-			while(!deck.StandardViable && !openDeck.IsWild)
-			{
-				var result = await this.ShowMessageAsync("Not a wild deck", "Open deck is a standard deck, but we are importing a wild deck. Please switch the deck to wild before continuing.", AffirmativeAndNegative, settings);
-				if(result == MessageDialogResult.Negative)
-					return;
-				openDeck = Reflection.GetEditedDeck();
-			}
-			var controller = await this.ShowProgressAsync("Creating Deck", "Please do not move your mouse or type.");
-			Topmost = false;
-			await Task.Delay(500);
-			var success = await DeckExporter.Export(deck, async () =>
-			{
-				if(controller != null)
-					await controller.CloseAsync();
-				ActivateWindow();
-				var result = await this.ShowMessageAsync("Importing interrupted", "Continue?", AffirmativeAndNegative,
-					new MessageDialogs.Settings() { AffirmativeButtonText = "Continue", NegativeButtonText = "Cancel" });
-				if(result == MessageDialogResult.Affirmative)
-					controller = await this.ShowProgressAsync("Creating Deck", "Please do not move your mouse or type.");
-				return result == MessageDialogResult.Affirmative;
-			});
-			if(controller.IsOpen)
-				await controller.CloseAsync();
-			if(success)
-			{
-				var hsDeck = Reflection.GetEditedDeck();
-				if(hsDeck != null)
-				{
-					var existingHsId = DeckList.Instance.Decks.Where(x => x.DeckId != deck.DeckId).FirstOrDefault(x => x.HsId == hsDeck.Id);
-					if(existingHsId != null)
-						existingHsId.HsId = 0;
-					deck.HsId = hsDeck.Id;
-					DeckList.Save();
-				}
-			}
+			DeckExportFlyout.Deck = deck;
+			FlyoutDeckExport.IsOpen = true;
 		}
 
 		public void ShowScreenshotFlyout()
@@ -249,7 +180,7 @@ namespace Hearthstone_Deck_Tracker.Windows
 			if(result.WasCancelled)
 				return;
 			if(result.Deck != null)
-				ExportDeck(result.Deck);
+				ShowExportFlyout(result.Deck);
 			else
 				await this.ShowMessageAsync("No deck found", "Could not find a deck on" + Environment.NewLine + result.Url);
 		}